+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Enable babel transpilation for released React components
 * Fix ESLint resolver issues
 * Updates to release process
+* Add `--sprite` option to the eject CLI function
 
 ## v0.0.0-alpha.1 [Commits](https://github.com/airtame/airtame-icons/commits/v0.0.0-alpha.1)